changes for NEWRUN: now all allocation routines except CopyCT checks for LVRA (if needed). Also added code for handling stackobjects in AOA
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@2642 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/NEWRUN/ExitObjects.c
@@ -43,7 +43,7 @@ long *ExO(long *jumpAdr,
     if ((long)this->Proto == (long)ComponentPTValue) {
       struct Component *comp = (struct Component *)this;
       struct Component *callerComp = comp->CallerComp;
-      printf("ExO: passing comp 0x%x\n", comp);
+      printf("ExO: passing comp 0x%x\n", (int)comp);
       SP   = (long *)callerComp->SPx;
       PC   = (long *)callerComp->CallerLSC;
       /* TerminateComponent: (see Attach.c) */
